+++ CHANGES.md
@@ -48,6 +48,10 @@ OpenSSL 4.0
 
    *Bob Beck*
 
+* The ASN1_STRING_FLAG_X509_TIME define has been removed.
+
+   *Bob Beck*
+
  * various function parameters have been constified,
    in particular for X509-related functions.